Migrated MemberServiceTests.
